move out of the performance folder
