+++ static/app/components/events/interfaces/crashContent/exception/util.tsx
@@ -1,0 +1,21 @@
+import type {Frame} from 'sentry/types';
+import {getFileExtension} from 'sentry/utils/fileExtension';
+
+const fileNameBlocklist = ['@webkit-masked-url'];
+export function isFrameFilenamePathlike(frame: Frame): boolean {
+  let filename = '';
+  try {
+    filename = new URL(frame.absPath ?? '').pathname.split('/').reverse()[0];
+  } catch {
+    // do nothing
+  }
+
+  return (
+    // If all filenames are anonymous, we do not want to show this alert
+    (frame.filename === '<anonymous>' && frame.inApp) ||
+    // If all function names are on the blocklist, we do not want to show this alert
+    fileNameBlocklist.includes(frame.function ?? '') ||
+    // If all absolute paths do not have a file extension, we do not want to show this alert
+    (!!frame.absPath && !getFileExtension(filename))
+  );
+}
